Fix problems when closing a tab.
There were problems when opening a tab and then closing the tab.
The remaining tabs would not operate properly: the zoom area,
the logarithmic scale check buttons and the point ordering area
would not become hidden when supposed to, which should be fixed
by removing the tab only after removing references to the
zoom area, logarithmic scale check buttons and the point
ordering area.